3PAR FC: add ability to add WWNs to host
When a host is created on the HP 3PAR array,
at initialize connection time, the host is
created with all WWNs provided in the connector
structure. However, the driver did not have
the ability to add new WWNs after a host is created.

This patch adds the ability to add WWNs to a
host, at initialize connection time, if a new
WWN is provided in the connector structure.

Change-Id: I6fd8a5511f83d5460da30ff5558a3e95964a95f5
Closes-bug: 1258229