Allow user to specify audit period.
The script 'cinder-volume-usage-audit' is generally run by cron to
generate volume and snapshot exists audit notifications that can be
consumed by ceilometer or other auditing systems. In this mode it sends
notifications for volumes that existed in the 'last completed audit
period' as calculated by cinder.utils.last_completed_audit_period()

However, it may sometimes be necessary to be able to send notifications
for audit periods other than than the last completed audit period. For
example if the host running the audit script was down for a period.

This patch provides three optional command line parameters to the script
cinder-volume-usage-audit. If no parameters are provided then the
operation of the script is unchanged. If parameters are provided then
they have the following effects:

--start_time specifies a date and time that is used in place of the
start of the last completed audit period.

--end_time specifies a date and time that is used in place of the end of
the last completed audit period.

--send_actions As well as sending exits notifications the script will
send any create or delete notifications that would have been sent during
the specified audit period.

Closes-Bug: #1259147

Change-Id: I0cfe963ea59d811ba4bac0bd3de31382a088b8c8