Add bool_type dictionary to test_migrations
A number of MySQL migration tests fail because the test case
is verifying that an sqlalchemy.types.BOOLEAN column was created.
When the column type is checked, however, mysql reports the type as
sqlalchemy.dialects.mysql.TINYINT.  The mismatch causes numerous
test cases to fail for MySQL.

This creates a dictionary for each engine being tested to handle the
different boolean types.

postgresql and DB2 encounter similar issues.  Subsequent patches
will make the appropriate changes for those databases.

Change-Id: I81cb3c6feeacf887f1f473ec54414e0e215d2c24
Closes-bug: 1260126