Remove legacy config option 'connection_type'
'connection_type' is a legacy config option originated from Nova.
Nova used this option to specify underlying virtualization type for
compute service and had deprecated it (with 'compute_driver') in
Folsom and had it removed in later release.  In Cinder Folsom release,
the XenSMDriver in Cinder used this config option to verify the target
hypervisor but in Grizzly XenSMDriver has been renamed and refactored
and no longer used this config.  Therefore, this option is dead for
a long time, and should be removed.

Change-Id: I4d72f31302fa813df42e71224151148a9c4a76c7