Revert initialize_connection changes
A change was merged that reusulted in interittent
gate failures.  It seems that in certain cases the
intitialize is doing an update targets and is
deleting the existing targets and not recreating them.

The idea of this patch was to fix up the iscis targets
for cases where volumes are extended.

A better solution is in progress that will remove the
target creation from the create_volume process altogether.

Currently we seem to have create_export, ensure_export and
now initialize that all are designed to do very similar
things.

A bug has been filed to address this and will attempt
to collapse these functions to be done at attach time.

This particular patch in essence reverts two patches:
commit: 18f8f5be94aae3b1747b143479ea1b188872f000
commit: a9267644ee09591e2d642d6c1204d94a9fdd8c82

Change-Id: Ib615847f3cbeb74b1e4416d51730162ad1d07a40
Closes-Bug: 1270608