Move clear_volume back to it's own method
This just moves clear_volume back to it's own method
and also only calls volutils.clear if we actually
need to perform the clear operation.

Also takes out the hard-coded multiplier of 1024 and
uses units.GiB.  Maybe this should be a separate patch,
I'll spin another if there are objections.

Change-Id: I56f6690fbadc0d44eba48f49d29386d5b9dca195