Update FibreChannel Zone Manager config
This patch changes the way the config options are
layed out in the cinder.conf to follow the
standards for grouping settings together.

Instead of using dynamic config names to group
settings together by fabric name, we look for
fabric groups based upon the fc_fabric_names.
We also create a new fc-zone-manager group
for all of the global fczm settings.

Change-Id: I91c8a05adec2fb12f664ca0df6007465c1a460e0
Closes-Bug: #1284338
DocImpact