Fix free_capacity reporting in SolidFire driver
The SolidFire driver reports capacity info in bytes, the
capabilities update reports available_GB.  Sadly we neglected
to convert the bytes to gigibytes here which made for a VERY
large backend and caused things like capacity filtering to not
work correctly.

This patch just adds conversion to GiB when reporting capabilities.

Change-Id: I62c6ad2edd8c2ced344df766c198504894f4902b
Closes-Bug: 1284452