GlusterFS: Increase snapshot delete job timeout to two hours
Increase the timeout for Nova snapshot delete operations from ten
minutes to two hours.  This helps prevent Cinder from terminating
operations prematurely that are still being processed by Nova.

It is not uncommon for snapshot delete jobs to run for longer than
ten minutes depending on the size of the snapshot and speed of the
storage backend.

This will be followed up with a more robust mechanism to keep track
of snapshot job progress as a later effort.

Related-Bug: 1273894

Change-Id: I1ad52568aed1ce1bf593e71704e481b6fe5f44fb