Can't force-create snapshot by an non-exist error volume
If we create an LVM volume whose size larger than
rest space in VG, the volume would not exist with
error status.
But then we still can force create a snapshot
by this volume, and the snapshot's status is available.
In the code,I found in create_lv_snapshot,
it return False when fail to get the volume.
However, raising an exception is only way to be
handle outside.
So, we should raise exception instead of return False.

Closes-Bug: #1283338

Change-Id: I80256f19d66da460c95ff23834abb79a557763bf