vmware: Mark VMware ESX vmdk driver as deprecated
The VMware ESX driver in Nova is now marked as deprecated and will be
removed in Juno release. This change marks the VMware ESX vmdk driver
as deprecated since it is intended for use with the ESX driver in Nova.
The plan is to remove the ESX vmdk driver in Juno release.

Closes-Bug: #1291194
DocImpact
Change-Id: I285f3461c1ccef56329b70417fb023405dfceb5e