vmware: Remove pbm_default_policy config option
The pbm_default_policy config option is used to associate a storage
policy with the volume backing if there is no associated volume_type
or the volume_type doesn't have the storage_profile option. A better
approach is to set the storage_profile option in the default_volume_type.
This change removes the pbm_default_policy and queries the policy
using the storage_profile option in the given volume_type.

Closes-Bug: #1291181
Closes-Bug: #1291804
DocImpact

Change-Id: I2519ec1145712ce6927dacde91db468d066af021