NetApp cmode iscsi: Fix QOS extra spec
This patch fixes the incorrect behavior where the NetApp cmode
iscsi driver will choose to create a volume on a flexVol that has
a QOS policy group that matches the specified QOS policy group
intended for the cinder volume. The correct behavior is to ignore
the QOS policy group of the flexVol and instead assign the QOS policy
group to the newly created cinder volume.

Change-Id: Ibeb8e3b965af8e79d0294c5cf97410da2b44d40c
Partial-Bug: #1288283