Import request_id middleware bug fix from oslo
There is a bug in request-id middleware that a subsequent API request
will overwrite a request-id of a previous request when multiple API
calls are processed in parallel in request_id middleware. The fix is
a drop-in replacement; cinder code does not need to be modified to
use the updated request_id middleware.

oslo change to fix the middleware:
d7bd9dc37ac3d6bc171cd2e290c772633ad20a32

Closes-Bug: #1289696
Change-Id: Ib211318c47681f841bcde04e490ccc2070bdf8de