Update config generator from OSLO
Cinders sample config is out of date with recent
updates needed for keystone, however we had an
old version of the oslo generator so there was
no way to pick these changes up locally by running
the generate tool.

This pulls the latest generator.py from oslo, and
we run the update script to get us back in sync.

Current head of oslo:
  commit 2eab986ef3c43f8d1e25065e3cbc1307860c25c7
  Merge: 2b08343 0dbcae1
  Author: Jenkins <jenkins@review.openstack.org>
  Date:   Wed Mar 26 14:47:07 2014 +0000

      Merge "Adds test condition in test_periodic"

Change-Id: Id43becfaa39d04153167e592a1233a5aef6016c9