Fixes cinder volume attach on Windows
The initiator name and the target name are inverted when passing
the arguments to the method which associates the iSCSI target to
an initiator. For this reason, this operation will fail.

Also, the connection to the iSCSI target cannot be initialized
properly as the method which gets portal information is missing
the return value.

Change-Id: I31034cd407d12d69a410f8c9e86a2e6821349dad
Closes-Bug: #1299131