Fix ISER scan retry option
The 'num_iscsi_scan_tries' option was renamed to
'num_volume_device_scan_tries'.

Remove it from the test where the deprecated name is
referenced since it is not used there anyway.

Fix the ISER driver to set the new config field rather than the
old one.  Presumably this made the num_iser_scan_tries option
not work since only the new field is passed to the brick
connector.

Closes-Bug: #1323741

Change-Id: I13ac4cce10894b541054fbda719bbc5fd555e3c1