vmware: Fixes create volume from image error
While creating a volume with an image specified as the volume source,
the VMDK driver tries creating the volume backing file on each host
in the inventory (one by one; until any attempt succeeds). However,
if the attempted host is in a bad connection state, the volume
creation fails and results in the volume going into error state.
Effectively, user cannot create a volume from image in such scenario.
This patch catches the VimFaultException during backing file
creation, so that driver picks the next available host for volume
creation.

Change-Id: I8ffdb849f41e8c28137ad8c0fa1f86c54b5e6fe9
Closes-Bug: #1326366