Add the -f option back to lvremove
Part of commit I4703133180567090878ea5047dd29d9f97ad85ab
add some logic to only use the -f flag on lvremove incase
of a failure condition during the retry.

The problem is that the initial call results in a prompt
pretty much every single time you call lvremove.  As such
that call fails every time and we go in to the retry with
force which succeeds.

This seems pretty wasteful, and there's no value that I can
see in continuing to run a cmd every time that we know is
going to fail.

This patch adds a -f back on to the first lvremove to eliminate
this problem.

Change-Id: I49a4882382f026cd286d0eda605a91cc7a574745
Closes-Bug: 1336811