Enable lazy translation for Cinder
This patch is the final step for getting lazy translation
re-enabled for Cinder.  It removes the use of gettextutils.install()
which is no longer needed with the addition of the explicit import
of _() in all of Cinder's files.  The configuration of 'cinder' catalog
is handled by gettextutils.

The install() function is replaced by gettextutils.enable_lazy() which
properly enables lazy translation when _() or _LX() is used.

Change-Id: I79e705f3852e8aa34034015199d56a8b23aeb8ef
Implements-Blueprint: i18n-enablement