Add iSCSI Target objects as independent objects
This patch is a step in decoupling the target
methods and the Volume Driver's Control methods.

This adds the targets directory and the new target objects
that we use with the exception of IET (follow up for that later).
TgtAdm and LIO drivers have been tested with the new LVM object.

All existing drivers are still able to be specified and use the
same objects and code-path they were using previously.  New
connector objects are only used when specifying the new driver.

Next step will be mapping current ref LVM driver to the new
LVM object and target model and working on the unit-tests.

After that mark the "old" methods and objects as deprecated
and we can then begin working on some other improvements.

Partial-Bug: #1329139

Change-Id: Iaa55e31e3dadc7dcb58112302c3807a8f92bcada