Fix KeyError exception in NetApp CDOT iscsi driver volume create.
Prior to commit 98aa91b, when the NetApp CDOT iscsi driver experiences
an NaAPIError interacting with a backend vserver, it catches (and logs)
the NaAPIError, but then throws an irrelevant KeyError because the Path
metadata field has not been set properly.  The KeyError obfuscates and
buries the real error message.

In the cinder pools implementation via commit 98aa91b this problem no
longer exists, but there is a need for a fix in codebase (e.g. icehouse)
where cinder pools will not be implemented.

Note that there are multiple ways that an NaAPIError can be thrown.  In
launchpad bug 1363763 an NaAPIError was thrown, but then buried behind a
KeyError exception, because of an attempt to create a cinder volume with
a Netapp qos policy-group when the backend container on the filer
already itself had a qos policy-group attached to it (outside of
Openstack).  The fix supplied here addresses all types of NaAPIError,
not just errors related to QOS.

Change-Id: Idc93ddddefdf27b7edf0813a14d8562fb5dcd175
Closes-Bug: 1363763