Netapp: fix multiple copies of cinder-volume
Various commands were failing when running multiple copies of
cinder-volume. This is due to storing an in memory copy of the luns
which may get updated in the backend by another cinder-volume. Fix
it by attempting to refresh the list of luns if one is not found
during a request.

Change-Id: Iecd1dc5146525d7240a09f71669a3ab239c488bd
Resolves-bug: #1367044