Fix NetAppDirectCmodeNfsDriver._find_shares()
If more than one NFS share is defined, all volumes are allocated on the
same NFS share until it is not eligible anymore. Then a second NFS share
is filled. The code is supposed to split volumes between the different
NFS shares, not fill them one by one.

It is a bug in _find_shares(). If at least one NFS share is eligible,
the "size" parameter (size of the requested volume in gigabytes) is
replaced with the total size in bytes of the share. Because of that,
_find_shares() cannot find more one eligible NFS share.

_find_shares() sorts eligible volumes by available size. Because of this
bug, all volumes are allocated in the same NFS share, instead of being
splitted on different NFS shares.

The _find_shares() has been removed in Juno, see:

* commit 98aa91b0e271d6fbd9dfe43bf4152fb138a46a89
* change Ie6f155df7bc1ae2cd5f7fa39f1b1a0ad38075988

Change-Id: I7b72921a2a45d5dd0aa969c8c32f7d815175f3a8
Co-author: "Florent Flament" <florent.flament-ext@cloudwatt.com>
Closes-Bug: #1369426