Fix ssh_host_key_file default in help and config.sample.conf
The commit message and the actual default say the default value for
ssh_host_key_file is $state_path/ssh_known_hosts, but the
config.conf.sample and the config opts help say it is
"$state_path/known_hosts".

Fix the help and config.conf.sample to match the actual default.

Closes-Bug: 1372765
Change-Id: I51a2918155a27afcf48eda01f32774ce0d346d06