Fix SolidFire inaccurate model on migrated vols
The general migration impl in Cinder works
by creating a new volume, transfering the data
from the original volume to the new volume, and
then deleting the original and flipping the ID
of the new volume.

Turns out we missed the fact that this creates a
mismatch between the volume Cinder will later ask
for and what the volumes identity is on the backend
device.

This change adds a check on create_volume at the drivers
level to see if it's part of a migration and is infact
going to get renamed.  If so, just use the new name
and avoid all the headaches that come later with updating
provider auth and location.

The model info won't change in this case and is accessible
independent of the ID field in the Cinder base and the
crazy change that's going to take place on that value
in the Cinder DB.

Change-Id: If916df5fd986352ffb20f2dd74552e4d7dad4f60
Closes-Bug: #1381943