Bump Req timeout to 30 seconds in SolidFire Driver
Local SolidFire CI system seems to have various networking
disconnects during runs.  This appears to be a problem with
latencies in our lab.

Currently have resources working on this, but in the meantime
bumping the timeout value in requests to see if we can address
it that way and get CI running again.  We'll either get to root
cause in our network or add this as a formal config option later
that will be used only for SolidFire CI system.

Change-Id: I0a13ac2e60c7053fc935922bfb28127ade858392