Update cinder.conf.sample to fix max db conn retries
Another external library change (oslo.db) has broken the gate
because cinder.conf.sample appears to be out of date.  The change
was just to a message string:

   Maximum number of database connection retries during
   startup. Set to -1 to specify an infinite retry count. (integer value)
   ->
   Maximum db connection retries during startup. Set to -1 to
   specify an infinite retry count. (integer value)

This patch updates cinder.conf.sample to get beyond this failure.  The
Cinder team will be talking about how to avoid these issues in the future.

Change-Id: Ia14954842d141b9bd208ce98c5a70a42b014ac45