Fix messages in EMC VMAX driver with no translation
There were a number of messages in the EMC VMAX driver
for LOG.info, LOG.error and LOG.warning that had no translation
marker on them.  This patch adds the appropriate _LI, _LE or _LW
marker so that the message will be translated.  Since there is
a separate effort to add in the log level markers going on I
did not attempt, in this patch, to address that issue.  Just
thought we should get translation on the messages that were missing it.

Change-Id: If6ac19369b303466afdec98589afa80ac46ad0f2