Fix _get_disk_of_partition edgecase in utils
_get_disk_of_partition in cinder/utils returns a disk device path,
calculated from a supplied partition device path, and stat for a device.

If the disk path calculated by the function, using a regex, is not a
block device the function will return the original device path argument
and the stat object for disk path that was not a block device.

This commit modifies the function so that if the supplied device path is
returned by the function the stat of that path is guaranteed to be
returned.

Change-Id: I7ce0a748fe4a0ebaa02cba11baadb8c71c007aa2