Fix find_autodoc_modules.sh to support OSX
When running 'find cinder/' on OSX platforms, all returned paths
will have 2 slashes (//). Because the script only strips
the first slash as per the CINDER_DIR value, we end up
with module names such as "cinder..db.api" in the documentation.

This change trims the leading dot if found to avoid this situation.

Change-Id: I9423b253d9842295850f469757e928bf5856967a