Fix 3PAR driver attach error when host name missing
In some cases an error was being thrown by the 3PAR driver
when the 'name' property returned from queryHost was
missing.  This change adds a check so that the driver
reacts properly when this occurs.  Also removed an unneeded
warning output that was still in the FC driver.

This patch is related to this merged bug fix:
https://review.openstack.org/#/c/138212/

Closes-Bug: 1398206
Change-Id: I6593dc68616f55496c42d625968adfc684f98032