Fix for typo in Purity Host create/delete methods in PureISCSIDriver
When refactoring the string manipulation methods there was a typo
introduced in:

https://git.openstack.org/cgit/openstack/cinder/commit/?
id=1ebe1bddd0a02e9efaa05c8ec9d71a299b6877e1

That change modified the methods to create/delete hosts in purity. The
REST API url was incorrectly pointing to /host/<host_name>/volume
instead of /host/<host_name>

This commit fixes the url and adds in unit tests to prevent this type of
regression in the future.

Change-Id: Iba24fc0a36e01e1d163b398bf32c41fbb9103861
Closes-Bug: 1399820