Fixing 3PAR connection name cache error
The 3PAR driver had a connection name cache in 3PAR common.  If
Cinder was restarted or the common object was destroyed the cache
was lost.  Instead of having the cache in 3PAR common, the 3PAR
driver can query the backend to see if there is a host already
available.  Replacing the name cache with a backend query had minimal
impact on performance of the driver and eliminates the problem of the
cache being destroyed/lost.

Closes-Bug: #1398914
Change-Id: Ie5467abc38946cceb4896fe9cbbd1606c7fbfba2