GET volumes API sorting REST/volume/DB updates
This change is to support updating the v2 /volumes and /volumes/detail
APIs to support multiple sort keys and sort directions.

Contains:
* Updates to the v2 REST API to pass multiple sort keys and directions
  to the volume API
* Volume API updates to accept and pass the sort information the DB layer
* API signature updates on the DB layer to accept a list of sort keys and
  directions

Note that the defaulting of the sort keys and directions is done in
the dependent patch set in the new 'process_sort_params' function
(invoked in db.sqlalchemy.api); by default, the sort keys are
'created_at' and 'id' in the 'desc' direction.

Partially Implements: blueprint cinder-pagination
APIImpact

Change-Id: I02bd8104b501b496148f95b0a5045fa2b64c8802