Fix HNAS driver confusing error message (iSCSI driver)
The error message shown when the parser finds a parser error
says, 'file not found' which causes confusion on the user when
he/she needs to debug the real cause of the problem. This patch fixes
this by testing first if the file exist and then throwing a proper error
message.

Also corrects an error when appending the config_file_name to the message on
the NFS driver and fixes LOG messages according to oslo.i18n guidelines.

Closes-Bug: #1403986
Change-Id: Ib6e613796eede751215db54696b353431caaafd1