Fix 3PAR host persona mapping to match WSAPI
The HP 3PAR drivers had mapped it's extra specs host personas to match
the 3PAR CLI/IMC, the WSAPI server has a different mapping for the
same personas. This patch corrects the mapping.

The Configuration Reference Guide will need to be updated to document
the correct mappings to match the code.
Closes-Bug: 1403997
DocImpact

Change-Id: I59272b4eb39392deb309771b10ebfecd0cf09454