Get volume from db again before updating it's status
During uploading a volume which is attached to an instance to image service,
if the instance is deleted then the volume remains in "in-use" status.
So get the volume again from db before updating it's status.

Added new db api called 'volume_update_status_based_on_attachment'.
It gets volume from db and updates its status using the same session.

Closes-Bug: #1406703
Change-Id: I01d313438f021cd2fe0d6f5c23f2029279d4b55a