Revert "Outputs the message about failing to bind
This reverts commit 8b59396182ceacb704e00afe1aee040a19e2a187.

This commit broke the ability to create Iscsi targets w/
cinder-rtstool. Everything seems to be working fine except there
are issues with regards to how rtstool output is parsed in that
any output (regardless of exit status) causes the command
to be marked as failed.

Suggesting this as a quick revert until we can investigate
the proper fix for error handling with cinder-rtstool.

Change-Id: I439433571ed1d01d162f685f8441b28bd6bffa43
Closes-bug: #1405228